The UD Update (#4)
The UDU (Upgraded Difficulty Update) is now out. Whats new in this update is; the controls for when a user chooses different difficulty setting, also it is no longer relying on speed to make it a difficult game as was suggested from a friend.

Along with this update comes NEW colors to the game based on the difficulty chosen.